+++ GraceLog/Source/Domain/Entity/DiaryKeyword.swift
@@ -1,0 +1,19 @@
+//
+//  DiaryKeyword.swift
+//  GraceLog
+//
+//  Created by 이건준 on 7/5/25.
+//
+
+import Foundation
+
+enum DiaryKeyword: String, CaseIterable {
+    case obedience    = "순종"
+    case faith        = "믿음"
+    case love         = "사랑"
+    case vision       = "비전"
+    case guidance     = "인내"
+    case peace        = "평안"
+    case suffering    = "고난"
+    case perseverance = "끈기"
+}
